fix: add setupMocks to setupTests
