+++ src/setupTests.ts
@@ -4,7 +4,24 @@
 // learn more: https://github.com/testing-library/jest-dom
 import "@testing-library/jest-dom";
 import "jest-canvas-mock";
-import { mockLogError, mockNavigation } from "./config/testUtils";
+import { initializeFirebase } from "./services";
 
-mockNavigation();
-mockLogError();
+export const mockNavigationFunction = jest.fn();
+export const mockLogErrorFunction = jest.fn();
+
+function setupMocks() {
+  jest.mock("hooks/useNavigation", () => ({
+    __esModule: true,
+    default: () => ({
+      navigateTo: mockNavigationFunction,
+      history: { location: {}, search: "" },
+    }),
+  }));
+  jest.mock("services/crashReport", () => ({
+    __esModule: true,
+    logError: mockLogErrorFunction,
+  }));
+}
+
+setupMocks();
+initializeFirebase();
